chore: renaming of files and folders
