+++ frontend/src/app/loading.tsx
@@ -1,4 +1,4 @@
-import FlowBackground from "@/components/Flow/FlowBackground/flowBackground";
+import FlowBackground from "@/components/Flow/Background/flowBackground";
 
 export default function Loading() {
   // You can add any UI inside Loading, including a Skeleton.
