+++ frontend/src/app/map/page.tsx
@@ -18,13 +18,13 @@ import { shallow } from "zustand/shallow";
 // we have to import the React Flow styles for it to work
 import Flashcard from "@/components/Flow/FlashcardComponents/Flashcard";
 import FlashcardEdge from "@/components/Flow/FlashcardComponents/FlashcardEdge";
-import useStore, { RFState } from "@/components/Flow/FlowElements/nodeStore";
+import useStore, { RFState } from "@/components/Flow/stores/nodeStore";
 import { Button } from "@/components/ui/button";
 import { useCallback, useRef, useState } from "react";
 import "reactflow/dist/style.css";
 
 // we need to import the React Flow styles to make it work
-import { FlowHeader } from "@/components/Flow/FlowHeader/FlowHeader";
+import { FlowHeader } from "@/components/Flow/Header/FlowHeader";
 import { useAddStack } from "@/components/Flow/addStacks";
 import "reactflow/dist/style.css";
 
@@ -149,7 +149,7 @@ function Map() {
         <MiniMap />
         <Controls showInteractive={false} />
         <Panel position="bottom-center" className="space-x-4">
-          <Button onClick={addStack}>Add Stack</Button>
+          <Button variant="default" onClick={addStack}>Add Stack</Button>
         </Panel>
       </ReactFlow>
     </div>
