+++ frontend/src/app/page.tsx
@@ -1,38 +1,95 @@
 // root page
+"use client";
+import FlowBackground from "@/components/Flow/Background/flowBackground";
+import { Card } from "@/components/ui/card";
+
 import { Button } from "@/components/ui/button";
+import { Dialog, DialogContent } from "@/components/ui/dialog";
+import { Box, Flex } from "@radix-ui/themes";
 import { ArrowRight } from "lucide-react";
 import Link from "next/link";
+import { useState } from "react";
+import { useStore } from "../components/Flow/stores/mapStore";
 
-/**
- * tailwind uses tree-shaking to optimize which classnames actually get injected in the css.
- * but we have some classes that we generate dynamically, which tailwind doesn't register.
- * to have them show up anyway, we inject them here.
- * ideally, we would use the `safelist` option in the tailwind config, but that doesn't seem to work.
- *
- * @see https://stackoverflow.com/questions/73660771/tailwind-safelist-patterns-for-multiple-patterns
- *
- */
-function TailwindClassPreloader() {
-  return (
-    <div className="bg-red-500 bg-green-500 bg-orange-500 bg-yellow-500 border-red-500 border-green-500 border-orange-500 border-yellow-500 hidden" />
-  );
-}
+function MapMenu() {
+  const [name, setName] = useState("");
+  const { addMap, categories, removeMap } = useStore();
 
-export default function Home() {
+  const sortedMaps = categories.sort((a, b) => {
+    return new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime();
+  });
+  const mapCards = sortedMaps.map((map) => (
+    <>
+      <Card style={{ maxWidth: 240 }} key={map.id}>
+        <Flex gap="3" align="center">
+          <Box>
+            <Link href={"/map"}>
+              <button>
+                <div className="p-3 rounded-lg m-10">{map.name}</div>
+              </button>
+            </Link>
+          </Box>
+        </Flex>
+      </Card>
+    </>
+  ));
+
+  const [openDialog, setOpenDialog] = useState(false);
+  const openNamingDialog = () => {
+    setOpenDialog(true);
+  };
+
   return (
-    <main className="bg-white">
-      <TailwindClassPreloader />
-      <div className="flex flex-col space-y-8 justify-center items-center h-screen">
-        <h1 className="text-3xl font-medium max-w-xl">
-          Welcome @rememberry 🫐
-        </h1>
-        <Link href={"/login"}>
-          <Button className="flex items-center">
-            <span>Go to the Login page</span>
-            <ArrowRight className="ml-2" />
-          </Button>
-        </Link>
+    <div className="relative">
+      <FlowBackground />
+      <div className="z-10 absolute top-0 left-0 w-full h-screen">
+        <Dialog open={openDialog}>
+          <DialogContent onAbort={() => setOpenDialog(false)}>
+            <div>
+              <div>
+                <h1 className="text-2xl font-medium max-w-xl text-blackberry">
+                  What do you want to call it?
+                </h1>
+                <input
+                  className=" rounded-md h-fit mt-4 my-2 bg-ashberry outline-none resize-none w-full break-words"
+                  placeholder="Map Name"
+                  value={name}
+                  onChange={(e) => setName(e.target.value)}
+                />
+              </div>
+
+              <div className="leading-6 text-justify"></div>
+              <Button
+                className="mt-4"
+                onClick={() => {
+                  addMap(name);
+                  setOpenDialog(false);
+                }}
+              >
+                Save
+              </Button>
+            </div>
+          </DialogContent>
+        </Dialog>
+        <div id="header"></div>
+        <div id="body" className="flex content-start p-20">
+          {mapCards}
+        </div>
+        <div
+          id="footer"
+          className="flex flex-row items-center justify-center space-x-4 fixed bottom-0 right-0 w-full h-16 bg-white"
+        >
+          <Link href={"/map"}>
+            <Button className="flex items-center ">
+              <span>Go to Map</span>
+              <ArrowRight className="ml-2" />
+            </Button>
+          </Link>
+          <Button onClick={openNamingDialog}>Add Map</Button>
+        </div>
       </div>
-    </main>
+    </div>
   );
 }
+
+export default MapMenu;
