+++ frontend/src/components/Authentication/CheckAuthentication.tsx
@@ -13,22 +13,22 @@ interface FetchUserProps {
 }
 
 export default function FetchUser({ user, children }: FetchUserProps) {
-  const router = useRouter();
-  const [isLoading, setIsLoading] = useState(true);
+  // const router = useRouter();
+  // const [isLoading, setIsLoading] = useState(true);
 
-  const userStore = useUserStore();
+  // const userStore = useUserStore();
 
-  useEffect(() => {
-    if (!user) {
-      router.push("/login");
-      userStore.actions.deleteUser();
-    } else {
-      userStore.actions.setUser(user);
-    }
-    setIsLoading(false);
-  }, [user, router, userStore.actions]);
+  // useEffect(() => {
+  //   if (!user) {
+  //     router.push("/login");
+  //     userStore.actions.deleteUser();
+  //   } else {
+  //     userStore.actions.setUser(user);
+  //   }
+  //   setIsLoading(false);
+  // }, [user, router, userStore.actions]);
 
-  if (isLoading) return <Loading />;
+  // if (isLoading) return <Loading />;
 
   return <>{children}</>;
 }
