+++ frontend/src/components/Flow/Header/FlowHeader.tsx
@@ -1,4 +1,4 @@
-import { DropDown } from "@/components/Flow/FlowHeader/DropDown";
+import { DropDown } from "@/components/Flow/Header/DropDown";
 import { MainStack } from "@/components/Flow/StackComponents/MainStack";
 import { ModeToggle } from "@/components/ui/ModeToggle";
 import { SidebarButton } from "@/components/ui/SidebarButton";
