+++ frontend/tailwind.config.ts
@@ -59,12 +59,28 @@ const config: Config = {
     secondary: "#ffffff",
   },
   colors: {
+    dark: {
+      900: '#121212', // very dark gray for backgrounds
+      800: '#1D1D1D', // for back panels or cards
+      700: '#2A2A2A', // for secondary containers or sidebars
+      600: '#383838', // for hover states or selected items
+      500: '#4D4D4D', // for borders and lines
+      400: '#626262', // for icons or secondary text
+      300: '#9E9E9E', // for placeholder text or disabled states
+      200: '#B3B3B3', // for input fields backgrounds
+      100: '#D4D4D4', // for text
+    },
     primary: "#1370DD",
     primarylight: "#5A97F1",
     secondary: "#ffffff",
-    ashberry: "#E6E9EF",
+    // dark colors
+    blackberry: "#282828",
+    dewberry: "333333",
     mulberry: "#C4C9D6",
-    blackberry: "#2E2E2E",
+    ashberry: "#E6E9EF",
+    snowberry: "#FFFFFF"
+    
+
   },
 
   plugins: [],
